requirements.txt
asgiref==3.7.2
asyncio==3.4.3
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-uuslug==2.0.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
uuid==1.30
whitenoise==6.5.0
